♻️ refactor: extract invitations api request
