+++ src/app/(with-header-sidebar)/dashboard/[id]/edit/_lib/invitationService.ts
@@ -1,0 +1,16 @@
+import axiosInstance from '@/lib/axiosInstance';
+import { GetInvitationsResponse } from '@/types/invitation';
+
+export const getInvitations = async (
+  id: string,
+  page: number = 1
+): Promise<GetInvitationsResponse> => {
+  try {
+    const response = await axiosInstance.get(
+      `/dashboards/${id}/invitations?page=${page}&size=5`
+    );
+    return response.data;
+  } catch (error) {
+    throw error;
+  }
+};
